test(pgregresstest): run pg_partman's pgTAP suite through multigateway
Add a pgTAP test harness alongside the pg_regress one so extensions that
ship pgTAP suites (pg_partman) run through multigateway. Each test file is
fed to psql and graded from the server-side TAP stream — no expected-output
files, no patch pipeline.

pg_partman is scoped to its transaction-wrapped tests (top-level test-*.sql
plus test_pg17plus/ and test_no_search_path/). The autocommit/procedure
subfolders are excluded and documented: pgTAP's plan() creates its state in
a temp table inside a function body, invisible to the gateway, so outside a
BEGIN..ROLLBACK it leaks across pooled backends ("you tried to plan twice").

Replace the CreateExtension bool with a schema-aware PreCreateExtensions
list shared by both harnesses (CreateExtension was the special case
{Name: ext.Name}); the pgTAP path tears its extensions down afterward so a
later suite (pgmq, which depends on pg_partman) starts clean.

Also splits the ~2k-line postgres_builder.go into pg_regress.go, pg_tap.go,
isolation.go, and report.go.
