+++ go/test/endtoend/pgregresstest/extensions.go
@@ -91,7 +91,7 @@ var ExtensionCatalog = []ExtensionInfo{
 	{"pg_graphql", KindExternal, StatusCovered, "Supabase pg_graphql; Rust/pgrx crate built with cargo-pgrx; loads test/fixtures.sql before its pg_regress suite"},
 	{"pg_jsonschema", KindExternal, StatusExternal, "Rust"},
 	{"pg_net", KindExternal, StatusExternal, "background worker"},
-	{"pg_partman", KindExternal, StatusUnsupported, "ships a pgTAP suite, not pg_regress; built and installed as a build dependency of pgmq (create_partitioned → create_parent)"},
+	{"pg_partman", KindExternal, StatusCovered, "pgTAP suite run via psql (not pg_regress); needs pgtap + max_locks_per_transaction>=128 (see testdata/pg17/external/pg_partman.conf). Runs the transaction-wrapped tests only (top-level + test_pg17plus/ + test_no_search_path/); autocommit/procedure subfolders can't run through a transaction pooler — see runExternalPgTAP. Also pgmq's build dependency (pgmq.create_partitioned → create_parent)."},
 	{"pg_stat_statements", KindContrib, StatusUnsupported, "NO_INSTALLCHECK; records query text the gateway rewrites"},
 	{"pg_trgm", KindContrib, StatusCovered, ""},
 	{"pgaudit", KindExternal, StatusExternal, ""},
@@ -99,7 +99,7 @@ var ExtensionCatalog = []ExtensionInfo{
 	{"pgjwt", KindExternal, StatusExternal, "depends on pgcrypto"},
 	{"pgmq", KindExternal, StatusCovered, "tembo-io/pgmq; pure-SQL queue built as a PGXS module from pgmq-extension/; partitioned-queue tests depend on pg_partman"},
 	{"pgsodium", KindExternal, StatusExternal, "libsodium"},
-	{"pgtap", KindExternal, StatusExternal, ""},
+	{"pgtap", KindExternal, StatusExternal, "test dependency of pg_partman; built from externalSpecs but not run as its own suite"},
 	{"plpgsql", KindContrib, StatusUnsupported, "built-in PL; exercised by the core regression suite, not contrib"},
 	{"plpgsql_check", KindExternal, StatusExternal, ""},
 	{"postgis", KindExternal, StatusExternal, ""},
@@ -110,6 +110,32 @@ var ExtensionCatalog = []ExtensionInfo{
 	{"uuid-ossp", KindContrib, StatusCovered, "needs --with-uuid"},
 	{"vector", KindExternal, StatusCovered, "pgvector; built as a PGXS module from externalSpecs"},
 	{"wrappers", KindExternal, StatusExternal, "Rust"},
+}
+
+// TestHarness selects how an external extension's shipped test suite is
+// executed. The zero value (HarnessPgRegress) drives the pg_regress binary and
+// diffs each test's output against expected/*.out — the model the contrib suite
+// and pgvector/pg_cron use. HarnessPgTAP instead feeds each test .sql to psql and
+// parses the TAP stream the pgTAP assertions emit server-side; correctness is
+// decided in-database (no expected-output files, no patch pipeline). Extensions
+// like pg_partman ship pgTAP suites.
+type TestHarness string
+
+const (
+	// HarnessPgRegress runs the extension's tests through pg_regress (default).
+	HarnessPgRegress TestHarness = ""
+	// HarnessPgTAP runs the extension's pgTAP tests through psql and parses TAP.
+	HarnessPgTAP TestHarness = "pgtap"
+)
+
+// ExtensionInstall names an extension to CREATE before a pgTAP suite, with an
+// optional target schema. When Schema is non-empty the harness creates that
+// schema first and installs the extension into it (CREATE EXTENSION ... SCHEMA
+// <Schema>); when empty the extension lands in the current schema (public). See
+// ExternalExtension.PreCreateExtensions.
+type ExtensionInstall struct {
+	Name   string
+	Schema string
 }
 
 // ExternalExtension describes one external (non-contrib) extension wired into
@@ -127,22 +153,63 @@ type ExternalExtension struct {
 	// pgmq-extension/, so it uses "pgmq-extension". Empty means the repo root.
 	BuildSubdir string
 
+	// Harness selects the test runner (see TestHarness). The zero value runs the
+	// pg_regress path; HarnessPgTAP runs the psql+TAP path. Fields below tagged
+	// "(pgTAP)" apply only to the HarnessPgTAP path; the patch pipeline applies only
+	// to the pg_regress path; PreCreateExtensions is used by both.
+	Harness TestHarness
+
+	// TestGlobs (pgTAP) are the filename globs, relative to TestSubdir, selecting
+	// the test files to run (their union, deduped). pgTAP suites ship flat *.sql
+	// files rather than the pg_regress sql/+expected/ layout, so listRegressTests
+	// does not apply. Defaults to ["*.sql"] when empty.
+	//
+	// pg_partman is restricted to its self-contained, transaction-wrapped tests:
+	// the top-level test-*.sql plus the rolled-back tests under test_pg17plus/ and
+	// test_no_search_path/. The OTHER subfolders are deliberately excluded, and the
+	// reason is a hard limit of running pgTAP through a transaction pooler, not a
+	// scoping whim — see the long note on runExternalPgTAP. In short: pgTAP keeps
+	// its plan/results in session-temp tables that plan() creates *inside* a
+	// function body, so the gateway can't observe them. Inside a BEGIN…ROLLBACK the
+	// visible BEGIN pins the backend and the ROLLBACK discards that temp state, so
+	// these tests are clean. The excluded folders run pgTAP in autocommit (their
+	// procedures COMMIT, so they can't be wrapped): plan()'s temp table is then
+	// created on an unpinned pooled backend and never discarded, leaking into the
+	// next file as "You tried to plan twice!". They also need infrastructure the
+	// pooled path can't provide — background workers (test_bgw/), tablespaces
+	// (test_tablespace/), non-superuser roles (test_nonsuperuser/), or manual
+	// multi-stage orchestration with out-of-band commits (test_procedure/).
+	TestGlobs []string
+
+	// ExcludeGlobs (pgTAP) removes files that TestGlobs would otherwise select,
+	// matched against the path relative to TestSubdir. Use it for
+	// a test that runs cleanly but isn't a reliable signal — e.g. a date-calibrated
+	// expectation that drifts with the calendar (see pg_partman below).
+	ExcludeGlobs []string
+
+	// PreCreateExtensions lists extensions to CREATE EXTENSION through multigateway,
+	// in order, before the suite runs — each optionally into a specific schema. Used
+	// by both harnesses for fixtures that assume an extension already exists:
+	//   - pg_regress: pgvector's fixtures open with a bare CREATE TABLE ...
+	//     vector(3) and never CREATE EXTENSION, so it lists {Name: "vector"}.
+	//   - pgTAP: pg_partman's test files never CREATE EXTENSION; they expect pgtap
+	//     in public and pg_partman in the `partman` schema, referencing partman.*
+	//     explicitly.
+	// The Schema field matters because pg_partman's control file is
+	// relocatable=false with no `schema=` default, so CREATE EXTENSION without a
+	// SCHEMA clause lands it in public (first in search_path) and every
+	// schema-qualified partman.* reference then fails with "schema partman does not
+	// exist". The pgTAP path tears these down after its suite (see runExternalPgTAP);
+	// the pg_regress path relies on resetContribState clearing public before the
+	// next extension.
+	PreCreateExtensions []ExtensionInstall
+
 	// TestSubdir is the directory within the checkout that holds the shipped
 	// pg_regress fixtures (sql/ + expected/), relative to the clone root.
 	// pgvector keeps them under test/; pg_cron keeps them at the repo root, so
 	// it uses "." (filepath.Join collapses it back to the clone root). pgmq keeps
 	// them under pgmq-extension/test, alongside its BuildSubdir.
 	TestSubdir string
-
-	// CreateExtension controls whether the harness pre-creates the extension
-	// through multigateway (and passes pg_regress --load-extension) before the
-	// suite runs. pgvector's fixtures assume it already exists (they open with a
-	// bare CREATE TABLE ... vector(3) and never CREATE EXTENSION), so it needs
-	// the preload. pg_cron's fixtures manage the extension themselves
-	// (CREATE EXTENSION pg_cron VERSION '1.0' is the first statement, then they
-	// DROP and recreate it at a newer version), so preloading would make that
-	// first statement fail with "extension already exists" — it must be false.
-	CreateExtension bool
 
 	// ScratchDatabases names databases the harness creates directly on the
 	// primary (bypassing multigateway, like the public-schema reset) before the
@@ -218,7 +285,9 @@ type ExternalExtension struct {
 var externalSpecs = map[string]ExternalExtension{
 	"vector": {
 		Name: "vector", Repo: "https://github.com/pgvector/pgvector", Tag: "v0.8.1",
-		TestSubdir: "test", CreateExtension: true,
+		// pgvector's fixtures assume the extension already exists (they open with a
+		// bare CREATE TABLE ... vector(3) and never CREATE EXTENSION), so preload it.
+		TestSubdir: "test", PreCreateExtensions: []ExtensionInstall{{Name: "vector"}},
 	},
 	"pg_graphql": {
 		Name: "pg_graphql", Repo: "https://github.com/supabase/pg_graphql", Tag: "v1.6.1",
@@ -228,9 +297,9 @@ var externalSpecs = map[string]ExternalExtension{
 		BuildSystem: "pgrx", PgrxVersion: "0.16.1", TestSubdir: "test",
 		// test/fixtures.sql opens with `drop extension if exists pg_graphql;
 		// create extension pg_graphql cascade;` and sets the graphql schema
-		// comment, so the harness loads it first and must not also preload the
-		// extension itself.
-		CreateExtension: false, FixturesFile: "fixtures.sql",
+		// comment, so the harness loads it first (FixturesFile) and must not also
+		// preload the extension itself (PreCreateExtensions left empty).
+		FixturesFile: "fixtures.sql",
 		// Several tests `create extension citext` — install it first (see
 		// ContribDeps), or they fail with "extension citext is not available".
 		ContribDeps: []string{"citext"},
@@ -241,20 +310,49 @@ var externalSpecs = map[string]ExternalExtension{
 	},
 	"pg_cron": {
 		Name: "pg_cron", Repo: "https://github.com/citusdata/pg_cron", Tag: "v1.6.4",
-		TestSubdir: ".", CreateExtension: false, ServerConfigFile: "pg_cron.conf",
+		// pg_cron's fixtures manage the extension themselves (CREATE EXTENSION pg_cron
+		// VERSION '1.0' is the first statement, then they DROP and recreate it at a
+		// newer version), so PreCreateExtensions is left empty to avoid colliding.
+		TestSubdir: ".", ServerConfigFile: "pg_cron.conf",
 		// pg_cron-test.sql references pgcron_dbno/pgcron_dbyes by name (it REVOKEs
 		// CONNECT on one and schedules/alters jobs targeting both) but never
 		// connects to them; front-load them on the primary so those metadata and
 		// CONNECT-privilege checks run for real. See ScratchDatabases.
 		ScratchDatabases: []string{"pgcron_dbno", "pgcron_dbyes"},
 	},
-	// pg_partman is a build-only dependency of pgmq, never tested on its own (it
-	// ships a pgTAP suite, not pg_regress — see its StatusUnsupported catalog
-	// entry). pgmq.create_partitioned calls partman's create_parent, which works
-	// without the background worker, so no ServerConfigFile is needed; the PGXS
-	// Makefile is at the repo root, so BuildSubdir stays empty.
+	// pgtap is a build dependency only: pg_partman's pgTAP tests need the pgtap
+	// extension installed. The harness clones/builds/installs it (via pg_partman's
+	// DependsOn) but never runs its own suite, so it carries no test fields and is
+	// not StatusCovered in the catalog.
+	"pgtap": {
+		Name: "pgtap", Repo: "https://github.com/theory/pgtap", Tag: "v1.3.4",
+	},
 	"pg_partman": {
 		Name: "pg_partman", Repo: "https://github.com/pgpartman/pg_partman", Tag: "v5.4.3",
+		Harness:    HarnessPgTAP,
+		TestSubdir: "test",
+		// The self-contained, transaction-wrapped tests only: the top-level
+		// test-*.sql plus the rolled-back tests under test_pg17plus/ and
+		// test_no_search_path/. The other subfolders are excluded — see TestGlobs.
+		TestGlobs: []string{"test-*.sql", "test_pg17plus/*.sql", "test_no_search_path/*.sql"},
+		// test-time-monthly-source-generated asserts an exact post-undo_partition
+		// row count (ARRAY[91]) calibrated to a specific run date: the data spans a
+		// fixed now()-relative 12-month window, but the monthly partition boundaries
+		// and premake shift with the calendar, so undo_partition(p_loop_count=>20)
+		// moves a date-dependent number of rows. It fails identically with and
+		// without the gateway (74≠91 on 2026-06-08) — the test's own date assumption,
+		// not a multigres behavior — so it's excluded from the deterministic set.
+		ExcludeGlobs: []string{"test_pg17plus/test-time-monthly-source-generated.sql"},
+		// pgtap (public) and pg_partman (partman schema) must both exist before any
+		// test file runs — the files assume them and never CREATE EXTENSION. pgtap
+		// goes in public; pg_partman MUST go in the `partman` schema (its tests
+		// reference partman.* explicitly), so it carries an explicit Schema.
+		DependsOn:           []string{"pgtap"},
+		PreCreateExtensions: []ExtensionInstall{{Name: "pgtap"}, {Name: "pg_partman", Schema: "partman"}},
+		// Subpartition tests create/drop several hundred tables in one transaction;
+		// the default max_locks_per_transaction (64) risks a cluster crash. pgmq,
+		// which DependsOn pg_partman, runs fine with this raised too.
+		ServerConfigFile: "pg_partman.conf",
 	},
 	"pgmq": {
 		Name: "pgmq", Repo: "https://github.com/tembo-io/pgmq", Tag: "v1.11.1",
@@ -263,8 +361,8 @@ var externalSpecs = map[string]ExternalExtension{
 		BuildSubdir: "pgmq-extension", TestSubdir: "pgmq-extension/test",
 		// Every test file CREATEs the extension itself (the topic/fifo files open
 		// with DROP EXTENSION IF EXISTS pgmq CASCADE; CREATE EXTENSION pgmq), so the
-		// harness must not preload it.
-		CreateExtension: false,
+		// harness must not preload it (PreCreateExtensions left empty).
+		//
 		// base.sql creates partitioned queues via pg_partman's create_parent and
 		// CREATEs pg_partman directly; install it first. (base.sql also calls
 		// pgmq.create_unlogged, whose CREATE UNLOGGED TABLE runs as dynamic SQL
