+++ go/test/endtoend/pgregresstest/pgregress_test.go
@@ -155,8 +155,8 @@ func TestPostgreSQLRegression(t *testing.T) {
 	// suite will run). Each is a PGXS module living outside the PostgreSQL source
 	// tree (e.g. pgvector); it is built against the just-installed PostgreSQL so
 	// its .so matches the running server's ABI. ExternalBuildList includes the
-	// covered extensions plus their build-only dependencies (e.g. pg_partman for
-	// pgmq), ordered so dependencies install first.
+	// covered extensions plus their build-only dependencies (e.g. pgtap for
+	// pg_partman, and pg_partman for pgmq), ordered so dependencies install first.
 	if runExternal {
 		t.Logf("Phase 2d: Installing external extensions...")
 		// Install contrib modules the external extensions depend on (e.g.
