+++ go/test/endtoend/pgregresstest/testdata/pg17/external/pg_partman.conf
@@ -1,0 +1,15 @@
+# pg_partman server configuration, appended at initdb time when the external
+# suite runs pg_partman (ExternalExtension.ServerConfigFile in extensions.go).
+#
+# pg_partman's pgTAP suite includes subpartitioning tests that create and drop
+# several hundred tables inside a single rolled-back transaction. The default
+# max_locks_per_transaction (64) is too small for that many relation locks in
+# one transaction and risks exhausting the shared lock table and crashing the
+# cluster, so pg_partman's own test README requires raising it; 128 is the value
+# its docs recommend. Appended after the pgctld template, so this wins.
+#
+# Note: pg_partman's background worker (pg_partman_bgw) is NOT loaded here. It is
+# only needed for automated maintenance and the test_bgw/ subfolder, which this
+# harness excludes (TestGlob = "test-*.sql"); CREATE EXTENSION pg_partman and all
+# top-level tests work without it.
+max_locks_per_transaction = 128
